requirements.txt
# CryptoTrader 依存パッケージ

# 基本ライブラリ
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2

# 取引所API
ccxt==4.1.40

# 機械学習
scikit-learn==1.3.2
lightgbm==4.1.0
hmmlearn==0.3.0

# データベース
# sqlite3は標準ライブラリ

# 通知
python-telegram-bot==20.6

# スケジューリング
APScheduler==3.10.4

# UI
streamlit==1.28.2
plotly==5.18.0
matplotlib==3.8.2

# 技術指標
pandas-ta==0.3.14b

# 設定ファイル
PyYAML==6.0.1
python-dotenv==1.0.0

# ユーティリティ
requests==2.31.0
tqdm==4.66.1

# AI/LLM
anthropic>=0.40.0

# バックアップ（オプション）
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0
# google-api-python-client==2.108.0

# 開発用
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
